feat: add LoginForm.jsx component for user authentication
